Project-AI_Career_Companion/backend/requirements.txt
azure-identity
azure-keyvault-secrets
opencensus-ext-azure
pytest
requests

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mlflow>=2.8.0
openai>=1.0.0
pydantic>=2.0.0

# Additional dependencies for enhanced logging and error handling
python-json-logger>=2.0.7
structlog>=23.2.0  # Optional: For structured logging
sentry-sdk[fastapi]>=1.38.0  # Optional: For error tracking